% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidDashboardBody.R
\name{fluidDashboardBody}
\alias{fluidDashboardBody}
\title{A dashboard body that automatically resizes when the sidebar is toggled}
\usage{
fluidDashboardBody(..., sidebarCollapsed = FALSE)
}
\arguments{
\item{...}{Items to put in the dashboard body.}

\item{sidebarCollapsed}{If \code{TRUE}, a javascript snippet will be inserted to collapse
the sidebar when the app is launched.}
}
\description{
\code{fluidDashboardBody} inserts a javascript snippet at the beginning of
the standard \code{\link{dashboardBody}} outuput, which tells the content of the dashboard
body to automatically adjust its sizing when the sidebar button is toggled.

It also provides the option to begin the app with the sidebar collapsed (\code{sidebarCollapsed = TRUE}).
}

