% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchExample.R
\name{launchExample}
\alias{launchExample}
\title{A launcher for example shiny apps}
\usage{
launchExample(name, ...)
}
\arguments{
\item{name}{A character string of the module (e.g. 'linkedScatter')}

\item{...}{Any extra arguments.}
}
\description{
\code{launchExample} provides an easy way to run examples
  of the modules supported in this package. These examples would ideally be
  located in the 'Help' documentation, but are too lengthy because they are
  shiny apps.
}
