% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScatter.R
\name{linkedScatter}
\alias{linkedScatter}
\alias{linkedScatterSidebarTabUI}
\alias{linkedScatterSidebarContentUI}
\alias{linkedScatterBodyUI}
\alias{linkedScatter}
\alias{scatterPlot}
\title{A pair scatter plots with linked brushing}
\usage{
linkedScatterSidebarTabUI(id, menu_item_name, tab_name)

linkedScatterSidebarContentUI(id, menu_item_name, tab_name, data)

linkedScatterBodyUI(id, tab_name)

linkedScatter(input, output, session, data)

scatterPlot(data, cols)
}
\arguments{
\item{id}{The module's unique identifier (to be passed to \code{\link[shiny]{NS()}})}

\item{menu_item_name}{Text to show for the menu item.}

\item{tab_name}{The name of a tab that this menu item will activate.}

\item{data}{A dataframe (or an object that can be correctly converted
to a dataframe by \code{as.data.frame()}) to be visualized as a scatterplot.}

\item{cols}{Variables ("columns") from the dataframe to be visualized in the scatter plot (the select input boxes
map to these parameters).}
}
\value{
Functions ending in 'UI' return a string of HTML. The output types of non-UI functions varies.
}
\description{
\code{linkedScatter} provides a set of functions to build a
linked scatter plot module. It is currently set up for use in a
\code{\link{shinydashboard}} framework.
}
