% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedBarMap.R, R/linkedScatterMap.R
\name{linkedScatterMap}
\alias{linkedBarMap}
\alias{linkedScatterMap}
\alias{linkedScatterMapBodyUI}
\alias{linkedScatterMapSidebarTabContentUI}
\alias{linkedScatterMapSidebarTabUI}
\title{A map with an interactive scatter plot}
\usage{
linkedScatterMapSidebarTabUI(id, menu_item_name, tab_name)

linkedScatterMapSidebarTabContentUI(id, menu_item_name, tab_name, sp)

linkedScatterMapBodyUI(id, tab_name)

linkedBarMap(input, output, session, sp_rx, plotly_event_rx)

linkedScatterMapSidebarTabUI(id, menu_item_name, tab_name)

linkedScatterMapSidebarTabContentUI(id, menu_item_name, tab_name, sp)

linkedScatterMapBodyUI(id, tab_name)

linkedScatterMap(input, output, session, sp_rx, plotly_event_rx)
}
\arguments{
\item{id}{The module's unique identifier (to be passed to \code{\link[shiny]{NS()}})}

\item{menu_item_name}{Text to show for the menu item.}

\item{tab_name}{The name of a tab that this menu item will activate.}

\item{data}{A dataframe (or an object that can be correctly converted
to a dataframe by \code{as.data.frame()}) to be visualized as a scatterplot.}

\item{cols}{Variables ("columns") from the dataframe to be visualized in the scatter plot (the select input boxes
map to these parameters).}

\item{id}{The module's unique identifier (to be passed to \code{\link[shiny]{NS()}})}

\item{menu_item_name}{Text to show for the menu item.}

\item{tab_name}{The name of a tab that this menu item will activate.}

\item{data}{A dataframe (or an object that can be correctly converted
to a dataframe by \code{as.data.frame()}) to be visualized as a scatterplot.}

\item{cols}{Variables ("columns") from the dataframe to be visualized in the scatter plot (the select input boxes
map to these parameters).}
}
\value{
Functions ending in 'UI' return a string of HTML. The output types of non-UI functions varies.

Functions ending in 'UI' return a string of HTML. The output types of non-UI functions varies.
}
\description{
\code{linkedScatterMap} provides a set of functions to build a
linked scatter plot-map module. It is currently set up for use in a
\code{\link{shinydashboard}} framework.

\code{linkedScatterMap} provides a set of functions to build a
linked scatter plot-map module. It is currently set up for use in a
\code{\link{shinydashboard}} framework.
}

