% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padded_div.R
\name{padded_div}
\alias{padded_div}
\title{HTML Helper: add padding to div element.}
\usage{
padded_div(..., pad_top = 0, pad_right = 20, pad_bottom = 20,
  pad_left = 20)
}
\arguments{
\item{...}{html contents of div.}

\item{pad_top}{top padding (pixels)}

\item{pad_right}{right padding (pixels)}

\item{pad_bottom}{bottom padding (pixels)}

\item{pad_left}{left}
}
\description{
Add padding to div element.
}
