% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCSSStyle.R
\name{shinyCSSStyle}
\alias{shinyCSSStyle}
\alias{checkboxInputStyle}
\alias{columnStyle}
\alias{sidebarCSS}
\title{Shiny UI function with 'style = ' arguments}
\usage{
checkboxInputStyle(inputId, label, value = FALSE, width = NULL,
  cssStyle = NULL)

columnStyle(width, ..., offset = 0, cssStyle = NULL)

sidebarCSS()
}
\arguments{
\item{inputId}{The module's unique identifier (to be passed to \code{\link[shiny]{NS()}}).}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value (TRUE or FALSE).}

\item{width}{The width of the input, e.g. '400px', or '100\%'.}

\item{cssStyle}{A string of style arguments using CSS syntax (e.g., 'padding:0px; margin-top: 5px;).}

\item{...}{Elements to include within the UI structure column (e.g. column).}

\item{offset}{The number of columns to offset this column from the end of the previous column.}
}
\value{
The UI element (as HTML) with any specified CSS styling.
}
\description{
\code{shinyCSSStyle} extend several common UI functions by adding a argument for css styling of a single element.
}
\details{
The \code{sidebarCSS} function added a lengthy set of CSS styling changes to the shinydashboard sidebar element.
}
