% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shmodules.R
\docType{package}
\name{shmodules}
\alias{shmodules}
\alias{shmodules-package}
\title{shmodules: a collection of shiny modules for data exploration and visualization}
\description{
A collection of shiny modules for exploring and visualizing data.
}
\details{
The primary purpose of this package is to make it easy to access successful
module patterns and quickly add them to new shiny projects.
}

