% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidDashboardBody.R
\name{fluidDashboardBody}
\alias{fluidDashboardBody}
\title{A dashboard body that automatically resizes when the sidebar is toggled}
\usage{
fluidDashboardBody(...)
}
\arguments{
\item{...}{Items to put in the dashboard body.}
}
\description{
\code{fluidDashboardBody} inserts a javascript "trigger "at the beginning of
the standard \code{\link{dashboardBody}} outuput, which tells the content of the dashboard
body to automatically adjust its sizing when the sidebar button is toggled.
}

