% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLeaflet.R
\name{myLflt}
\alias{myLflt}
\title{A leaflet map with Mapbox provider tiles}
\usage{
myLflt()
}
\value{
A leaflet map with provider tiles from Mapbox already loaded
}
\description{
\code{myLflt} provides a colorful basemap that can be used as the first argument of a Leaflet pipe chain.
}

